requirements.txt
numpy>=1.21.0
scipy>=1.9.0
pygame>=2.5.0
PyOpenGL>=3.1.0
PyOpenGL_accelerate>=3.1.0
websockets>=12.0
pyserial>=3.5
